backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Dotenv==0.0.3
python-dotenv==1.0.0
scikit-learn==1.4.2
pandas==2.2.0
numpy==1.24.3
joblib==1.3.2
